backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai==1.3.8
anthropic==0.7.8
pydantic==2.5.0
python-multipart==0.0.6
websockets==12.0
aiofiles==23.2.1
redis==5.0.1
chromadb==0.4.18
docker==6.1.3
sqlalchemy==2.0.23
asyncpg==0.29.0
pgvector==0.2.3
alembic==1.13.1